Bump x11-wm/mutter to unstable & add vrr support
Bump Mutter up to the current (unstable) point release, and add the
option to fold in the (currently in-progress, not-yet-merged-to-master)
variable refresh rate under Wayland patchset.  Shouldn't require any
functionality present in Clutter's git master (ahead of a stable point
release), but that's yet to be tested in the wild.